+++ src/components/App/App.js
@@ -1,43 +1,24 @@
 import agent from '../../agent';
-import Header from '../Header/Header';
 import React from 'react';
 import { connect } from 'react-redux';
-import { APP_LOAD, REDIRECT } from '../../constants/actionTypes';
+import { APP_LOAD } from '../../constants/actionTypes';
 import { Route, Switch } from 'react-router-dom';
 import Article from '../Article';
 import Editor from '../Editor/Editor';
-import Home from '../Home';
+import Home from '../../pages/home';
 import Login from '../Login/Login';
-import Profile from '../Profile/Profile';
-import ProfileFavorites from '../ProfileFavorites/ProfileFavorites';
+import Profile from '../../pages/profile';
 import Register from '../Register/Register';
 import Settings from '../Settings/Settings';
-import { store } from '../../store';
-import { push } from 'react-router-redux';
 
-const mapStateToProps = (state) => {
-  return {
-    appLoaded: state.common.appLoaded,
-    currentUser: state.common.currentUser,
-    redirectTo: state.common.redirectTo,
-  };
-};
+import MainWrapper from '../MainWrapper/MainWrapper';
 
 const mapDispatchToProps = (dispatch) => ({
-  onLoad: (payload, token) => dispatch({ type: APP_LOAD, payload, token, skipTracking: true }),
-  onRedirect: () => dispatch({ type: REDIRECT }),
+  onLoad: (payload, token) => dispatch({ type: APP_LOAD, payload, token, skipTracking: true })
 });
 
 class App extends React.Component {
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.redirectTo) {
-      // this.context.router.replace(nextProps.redirectTo);
-      store.dispatch(push(nextProps.redirectTo));
-      this.props.onRedirect();
-    }
-  }
-
-  componentWillMount() {
+  componentDidMount() {
     const token = window.localStorage.getItem('jwt');
     if (token) {
       agent.setToken(token);
@@ -47,34 +28,22 @@ class App extends React.Component {
   }
 
   render() {
-    if (this.props.appLoaded) {
-      return (
-        <>
-          <Header currentUser={this.props.currentUser} />
-          <Switch>
-            <Route exact path='/' component={Home} />
-            <Route path='/login' component={Login} />
-            <Route path='/register' component={Register} />
-            <Route path='/editor/:slug' component={Editor} />
-            <Route path='/editor' component={Editor} />
-            <Route path='/article/:id' component={Article} />
-            <Route path='/settings' component={Settings} />
-            <Route path='/@:username/favorites' component={ProfileFavorites} />
-            <Route path='/@:username' component={Profile} />
-          </Switch>
-        </>
-      );
-    }
     return (
-      <>
-        <Header currentUser={this.props.currentUser} />
-      </>
+      <MainWrapper>
+        <Switch>
+          <Route exact path='/' component={Home} />
+          <Route path='/@:username/favorites' component={Profile} />
+          <Route path='/@:username' component={Profile} />
+          <Route path='/login' component={Login} />
+          <Route path='/register' component={Register} />
+          <Route path='/editor/:slug' component={Editor} />
+          <Route path='/editor' component={Editor} />
+          <Route path='/article/:id' component={Article} />
+          <Route path='/settings' component={Settings} />
+        </Switch>
+      </MainWrapper>
     );
   }
 }
 
-// App.contextTypes = {
-//   router: PropTypes.object.isRequired
-// };
-
-export default connect(mapStateToProps, mapDispatchToProps)(App);
+export default connect(null, mapDispatchToProps)(App);
