+++ src/constants/actionTypes.js
@@ -1,5 +1,4 @@
 export const APP_LOAD = 'APP_LOAD';
-export const REDIRECT = 'REDIRECT';
 export const ARTICLE_SUBMITTED = 'ARTICLE_SUBMITTED';
 export const SETTINGS_SAVED = 'SETTINGS_SAVED';
 export const DELETE_ARTICLE = 'DELETE_ARTICLE';
@@ -32,5 +31,3 @@ export const UPDATE_FIELD_AUTH = 'UPDATE_FIELD_AUTH';
 export const UPDATE_FIELD_EDITOR = 'UPDATE_FIELD_EDITOR';
 export const FOLLOW_USER = 'FOLLOW_USER';
 export const UNFOLLOW_USER = 'UNFOLLOW_USER';
-export const PROFILE_FAVORITES_PAGE_UNLOADED = 'PROFILE_FAVORITES_PAGE_UNLOADED';
-export const PROFILE_FAVORITES_PAGE_LOADED = 'PROFILE_FAVORITES_PAGE_LOADED';
