+++ src/reducers/articleList.js
@@ -7,9 +7,7 @@ import {
   HOME_PAGE_UNLOADED,
   CHANGE_TAB,
   PROFILE_PAGE_LOADED,
-  PROFILE_PAGE_UNLOADED,
-  PROFILE_FAVORITES_PAGE_LOADED,
-  PROFILE_FAVORITES_PAGE_UNLOADED,
+  PROFILE_PAGE_UNLOADED
 } from '../constants/actionTypes';
 
 export default ( state = {}, action ) => {
@@ -69,7 +67,6 @@ export default ( state = {}, action ) => {
         tag: null,
       };
     case PROFILE_PAGE_LOADED:
-    case PROFILE_FAVORITES_PAGE_LOADED:
       return {
         ...state,
         pager: action.pager,
@@ -79,7 +76,6 @@ export default ( state = {}, action ) => {
         currentPage: 0,
       };
     case PROFILE_PAGE_UNLOADED:
-    case PROFILE_FAVORITES_PAGE_UNLOADED:
       return {};
     default:
       return state;
